requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-login==0.5.0
flask-cors==3.0.10
openai==1.78.1
python-dotenv==0.19.0
werkzeug==2.0.1
gunicorn==20.1.0
sqlalchemy==1.4.23 